functions/summary/requirements.txt
networkx==2.1
newspaper3k==0.2.8
nltk==3.4
